MDEV-31949 parallel slave xa Round-Robin distribution
XA-Prepare group of events

  XA START xid
  ...
  XA END xid
  XA PREPARE xid

and its XA-"complete" terminator

  XA COMMIT or
  XA ROLLBACK

are made distributed Round-Robin across slave parallel workers.
The former hash-based policy was proven to attribute to execution
latency through creating a big - many times larger than the size
of the worker pool - queue of binlog-ordered transactions
to commit.

Acronyms and notations used below:

  XAP := XA-Prepare event or the whole prepared XA group of events
  XAC := XA-"complete", which is a solitary group of events
  |W| := the size of the slave worker pool
  Subscripts like `_k' denote order in a corresponding sequence
     (e.g binlog file).

KEY CHANGES:

The parallel slave
------------------
driver thread now maintains a list XAP:s currently
in processing. Its purpose is to avoid uncontrolled parallel execution
of XAP and XAC of the same transaction as well as XA:s
with duplicate xids (unlikely, but that's the user's right).
The sliding window helps to identify any pair of XAP_k and XAC_k+i
whose distance between each other is less than |W| and therefore
the XAC part can't optimistically scheduled without a risk it
won't find its xid in the system cache.

To explain who it works in the following example let |W| the # of
Workers be 4.
Let Q_i be worker queues, that contain transactions presented
as their sequence numbers. They assigned to workers left to right, and
when the current row ends the next row is allocated for placing onto it.
Let there are just two rows of transactions in the queues:

  Q1 ...Q4
  1 2 3 4
  5 6 7 8

A maximum compelling to wait distance pair like
  XAP_1(xid) -> XAC_4(xid)
would enforce XAC_4 to waiting at the optimistic parallel mode
scheduling.
The XAC (XA-"COMPLETE" query event) would be allowed to begin
its execution only after T_3 has signaled its binlogging is done.
Note that the dependency ceases for the 6:th position and up:
   XAP_1 --not-> XAC_6 etc
That is XAC_6 could be sent into execution by that worker without
waiting.

If it's XAC_4 that has received the binlog ordering signal and has
begun its execution, it is still not guaranteed that it will
acquire the xid.
The latter is made possible by a developed mechanism of passing xids
from XAP:s that are completing (slave_applier_reset_xa_trans())
to XAC:s that are potentially waiting (xid_cache_search() slave extension).

There also exists a concern of the duplicate xid, which is depicted as
      XAC_i -> XAP_i+k
dependency on xid.
XAP_i+k is allowed to start early and race to the xid even before
XAC_i, and then only to be ER_XAER_DUPID errored out for optimistic
parallel mode retry.

XA related
----------
XID_cache_element is extended with a pointer to THD::rgi_slave
object of a XAC-waiter. The waitee and waiter coordinate xid
passing to each other, which a pre-existing pair of

   rpl_group_info::sleep_mutex,cond

helps to conduct.
The parallel slave worker has to execute
   xid_cache_search_maybe_wait()
instead of the normal xid_cache_search().

Replication performance with the slave binlog OFF:
-------------------------------------------------
Is improved with relocation of the commit order condition
from a before-engine-prepare (sql/log_event_server.cc) to
an after engine-prepare point of XA-PREPARE execution (unlog_xa_prepare()).
That exploits a group-prepare in Engine.

TODO/Remove: storage/innobase/trx/trx0undo.cc changes are due to
             a fixed MDEV-32144 (the fixes may not be yet in the XA
	     branch).